# YJIT: Implement codegen for Kernel#block_given?