@XrXr 2023-01-31
I think this is a semantic change, no? Now users can override `block_given?` on their class and change `Kernel#loop`'s behavior.