@k0kubun 2023-01-31
We (not me) have occasionally replaced rb_funcall with an equivalent C API as an "optimization". The opposite would be also accepted as long as it doesn't break existing sane or popular implementation. If you break block_given?, many other places could break too, so I don't think that's a valid concern.